Move ImGuiContext from imgui to imgui.internal, minor bugfixes
